requirements.txt
astroid==2.3.3
certifi==2019.11.28
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pylint==2.4.4
six==1.14.0
wincertstore==0.2
wrapt==1.11.2
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
colorama==0.4.3
future==0.18.2
imutils==0.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
tkintertable==1.3.2
wincertstore==0.2
wrapt==1.11.2
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
colorama==0.4.3
future==0.18.2
imutils==0.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
tkintertable==1.3.2
wincertstore==0.2
wrapt==1.11.2
